requirements.txt
Flask==3.1.2
flask_wtf==1.2.2
injector==0.23.0
openai==2.14.0
python-dotenv==1.2.1
WTForms==3.2.1
pytest
flask-sqlalchemy
sqlalchemy